Extract ListenerContainer interface, make implementation internal
